Add demo test case for https://hibernate.atlassian.net/browse/HHH-18502
Run the 2 tests in org.hibernate.bugs.JPAUnitTestCase

Observe one test fails and one test passes. The only difference is that
one test passes a Collection to the SqmSubQuery#in method, the other
method passes an array.

The error for the failing test is

> org.hibernate.HibernateException: Could not convert 'java.util.ImmutableCollections$List12' to 'java.lang.String' using 'org.hibernate.type.descriptor.java.StringJavaType' to wrap

As explained in the issue ticket, the problem is that SqmSubQuery#in calls the
wrong overload from the NodeBuilder.